+++ src/CustomerOrdersReport.php
@@ -5,14 +5,19 @@ namespace SilverCommerce\Reports;
 use DateTime;
 use SilverStripe\ORM\DB;
 use SilverStripe\ORM\ArrayList;
+use SilverStripe\Reports\Report;
+use SilverStripe\Forms\DateField;
+use SilverStripe\Forms\FieldList;
 use SilverStripe\Control\HTTPRequest;
-use SilverCommerce\Reports\SalesReport;
 use SilverStripe\Core\Injector\Injector;
 use SilverCommerce\OrdersAdmin\Model\Invoice;
 use SilverCommerce\ContactAdmin\Model\Contact;
 use SilverCommerce\Reports\Items\CustomerOrdersItem;
+use SilverStripe\Forms\GridField\GridFieldSortableHeader;
+use SilverStripe\Security\Member;
+use Symbiote\GridFieldExtensions\GridFieldTitleHeader;
 
-class CustomerOrdersReport extends SalesReport
+class CustomerOrdersReport extends Report
 {
     public function title()
     {
@@ -26,7 +31,21 @@ class CustomerOrdersReport extends SalesReport
     {
         return _t(
             __CLASS__ . ".Description",
-            "View a summary of customer orders"
+            "View a summary of customer orders within the time period"
+        );
+    }
+
+    public function columns()
+    {
+        $singleton = Contact::singleton();
+
+        return array(
+            'FirstName' => _t(CustomerOrdersItem::class . 'FirstName', 'First Name'),
+            'Surname' => _t(CustomerOrdersItem::class . 'Surname', 'Surname'),
+            'Email' => _t(CustomerOrdersItem::class . 'Email', 'Email'),
+            'InvoiceCount' => _t(CustomerOrdersItem::class . 'InvoiceCount', 'Total Invoices'),
+            'InvoiceTotal' => _t(CustomerOrdersItem::class . 'InvoiceTotal', 'Total Invoice Value'),
+            'InvoiceTax' => _t(CustomerOrdersItem::class . 'InvoiceTax', 'Total Invoice Tax')
         );
     }
 
@@ -37,7 +56,6 @@ class CustomerOrdersReport extends SalesReport
         $db = DB::get_conn();
         $start = null;
         $end = null;
-        $date_filter = null;
         $list = ArrayList::create();
 
         if ($params && array_key_exists("StartDate", $params)) {
@@ -54,55 +72,61 @@ class CustomerOrdersReport extends SalesReport
         }
 
         $format = "%Y-%m-%d";
-        $start_field = $db->formattedDatetimeClause(
+        $field = $db->formattedDatetimeClause(
             '"Estimate"."StartDate"',
             $format
         );
+        $date_filter = [
+            $field . ' <= ?' =>  $end->format("Y-m-d"),
+            $field . ' >= ?' =>  $start->format("Y-m-d")
+        ];
 
-        if ($start && $end) {
-            $date_filter = [
-                $start_field . ' <= ?' =>  $end->format("Y-m-d"),
-                $start_field . ' >= ?' =>  $start->format("Y-m-d")
-            ];
+        $invoices = Invoice::get()->where($date_filter);
+        $ids = $invoices->column('CustomerID');
+
+        // If no invoices available, set a null value (to return an empty list)
+        if (count($ids) === 0) {
+            $ids = 0;
         }
 
-        $contacts = Contact::get();
-        $orders = Invoice::get();
+        $contacts = Contact::get()->filter('ID', $ids);
 
-        if ($date_filter) {
-            $orders = $orders->where($date_filter);
-        }
-
+        // Loop through relevent contacts and create/assign relevent objects
         foreach ($contacts as $contact) {
-            $item = CustomerOrdersItem::create();
-            $item->Surname = $contact->Surname;
-            $item->Email = $contact->Email;
-            $item->PhoneNumber = $contact->Phone;
-            $total = 0;
-            $c_orders = $orders->filter('CustomerID', $contact->ID);
-            if ($c_orders->exists()) {
-                foreach ($c_orders as $order) {
-                    $total += $order->getTotal();
-                }
-            }
-            $item->TotalOrderValue = $total;
+            $item = CustomerOrdersItem::create()
+                ->setContact($contact)
+                ->setStartDate($start)
+                ->setEndDate($end);
 
             $list->add($item);
         }
 
-        return $list;
+        return $list->sort("InvoiceTotal DESC");
     }
 
-    // which fields on that object we want to show
-    public function columns() 
+    public function parameterFields()
     {
-        $fields = [
-            "Surname" => "Surname",
-            "Email" => "Email",
-            "PhoneNumber" => "PhoneNumber",
-            "TotalOrderValue" => "TotalOrderValue"
-        ];
+        $fields = FieldList::create();
+
+        $fields->push(DateField::create("StartDate", "Start"));
+        $fields->push(DateField::create("EndDate", "End"));
 
         return $fields;
+    }
+
+    /**
+     * Return the parent GridField without column sorting
+     *
+     * @return \SilverStripe\Forms\FormField subclass
+     */
+    public function getReportField()
+    {
+        /** @var \SilverStripe\Forms\GridField\GridField */
+        $grid_field = parent::getReportField();
+        $grid_field->getConfig()
+            ->removeComponentsByType(GridFieldSortableHeader::class)
+            ->addComponent(new GridFieldTitleHeader());
+
+        return $grid_field;
     }
 }